# Huge asymptotic speedup for base conversion